Reconciling the randomizer.
... the first of two rounds.
This will introduce variability of output in the 1991 translation, which arises, in part, from the use of srand(). 
Therefore, ‟make test” will not always produce an exact match for test outputs anymore.
This will be fixed for both C translations and the Fortran version later, when they are fully reconciled.