Incremental upgrade ... message output routine II.
... optimizations that account for more regularities that were originally present in the 1991 C translation, which the translator failed to notice.
The rspeak2() routine is actually a tree-walking routine in disguise; its true nature obscured by the fact that it is limited to just one level.
The "top" flag is actually the level counter, in disguise.
The variable "w" (formerly "iloc") is actually the value stack for the file location variable "x" - likewise limited to one level.
The variable "x" is equal to the current location in the story file, itself (i.e. x == ftell(StoryF)).